FPLA-1223: Converted webpack to ts
