+++ Dockerfile
@@ -1,8 +1,8 @@
 # ---- Base image ----
 FROM hmctspublic.azurecr.io/base/node:12-alpine as base
 RUN yarn config set proxy "$http_proxy" && yarn config set https-proxy "$https_proxy"
-COPY package.json yarn.lock webpack.config.js ./
-RUN yarn install --production  && yarn cache clean
+COPY package.json yarn.lock webpack.config.ts ./
+RUN yarn install --production && yarn cache clean
 
 # ---- Build image ----
 FROM base as build
