+++ package.json
@@ -6,11 +6,11 @@
     "node": ">=12.0.0"
   },
   "scripts": {
-    "start": "./node_modules/.bin/ts-node -r tsconfig-paths/register src/main/server.ts",
+    "start": "concurrently \"yarn run server\" \"yarn webpack-dev-server\"",
     "cichecks": "yarn && yarn build && yarn lint && yarn test && yarn test:routes && yarn test:a11y",
     "lint": "sass-lint -v -q && eslint . --ext .js,.jsx,.ts,.tsx ",
     "build": "yarn webpack",
-    "build:dev": "concurrently \"yarn run start\" \"yarn webpack-dev-server\"",
+    "server": "./node_modules/.bin/ts-node -r tsconfig-paths/register src/main/server.ts",
     "build:prod": "yarn webpack --mode production --config webpack.config.js",
     "test": "TS_NODE_TRANSPILE_ONLY=true NODE_ENV=mocha LOG_LEVEL=OFF mocha --opts mocha.opts $(find src/test \\( -name '*.ts' \\) ! -path '*a11y*' ! -path '*/routes/*') --reporter-options reportFilename=unit,inlineAssets=true,reportTitle=adoption-frontend-unit",
     "test:routes": "TS_NODE_TRANSPILE_ONLY=true NODE_ENV=mocha LOG_LEVEL=OFF mocha --opts mocha.opts $(find src/test \\( -name '*.ts' \\) -path '*/routes/*') --timeout 10000 --reporter-options reportFilename=routes,inlineAssets=true,reportTitle=adoption-frontend-routes",
@@ -31,6 +31,7 @@
     "@types/config": "^0.0.36",
     "@types/cookie-parser": "^1.4.2",
     "@types/cookies": "^0.7.4",
+    "@types/copy-webpack-plugin": "^5.0.0",
     "@types/csurf": "^1.9.36",
     "@types/es6-promisify": "^6.0.0",
     "@types/express": "^4.17.2",
@@ -38,6 +39,7 @@
     "@types/i18next-node-fs-backend": "^2.1.0",
     "@types/i18next-sprintf-postprocessor": "^0.2.0",
     "@types/jsonwebtoken": "^8.3.7",
+    "@types/mini-css-extract-plugin": "^0.9.1",
     "@types/moment": "^2.13.0",
     "@types/nock": "^11.1.0",
     "@types/node": "^13.7.0",
