--- src/main/assets/scripts/main.js
@@ -1,4 +1,0 @@
-// Adoption JS
-export const initAdoptionModules = () => {
-  // Initalise modules here
-};
