+++ src/main/assets/scripts/main.ts
@@ -1,0 +1,4 @@
+// Adoption JS
+export const initAdoptionModules = (): void => {
+  // Initalise modules here
+};
